MAINT: Refactor authentication dependencies
