+++ app/api/dependencies/authentication.py
@@ -1,15 +1,66 @@
 from datetime import datetime, timedelta
 from typing import Optional
 
-from jose import jwt
+from fastapi import Depends, HTTPException, status
+from jose import JWTError, jwt
 from sqlalchemy.orm import Session
 
+from app.api.dependencies.database import get_db
 from app.core.config import TOKEN_EXPIRE, ALGORITHM, SECRET_KEY
 from app.db.queries import users
 from app.models.schemas.users import UserInDB
+from app.services.security import oauth2_scheme
 
 
-def get_user_by_username(
+def authenticate_user(
+    db: Session, username: str, password: str
+) -> Optional[UserInDB]:
+    """ Get user data that is matched username and password
+    """
+    user = _get_user_by_username(db, username)
+    if not user:
+        return None
+    if not user.verify_password(password):
+        return None
+    return user
+
+
+def create_access_token(data: dict[str, object]) -> str:
+    expires_delta = timedelta(minutes=TOKEN_EXPIRE)
+
+    to_encode = data.copy()
+    expire = datetime.utcnow() + expires_delta
+
+    to_encode.update({"exp": expire})
+    encode_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
+    return encode_jwt
+
+
+async def get_current_user(
+    token: str = Depends(oauth2_scheme),
+    db: Session = Depends(get_db)
+) -> UserInDB:
+    credentials_exception = HTTPException(
+        status_code=status.HTTP_401_UNAUTHORIZED,
+        detail="Couldn't invalid credentials",
+        headers={"WWW-Authenticate": "Bearer"},
+    )
+    try:
+        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
+        username: str = payload.get("sub")
+        if username is None:
+            raise credentials_exception
+    except JWTError:
+        raise credentials_exception
+
+    user = _get_user_by_username(db, username=username)
+    if user is None:
+        raise credentials_exception
+
+    return user
+
+
+def _get_user_by_username(
     db: Session, username: str
 ) -> Optional[UserInDB]:
     db_user = users.get_user_by_username(db, username)
@@ -23,14 +74,3 @@ def get_user_by_username(
         )
 
     return None
-
-
-def create_access_token(data: dict[str, object]) -> str:
-    expires_delta = timedelta(minutes=TOKEN_EXPIRE)
-
-    to_encode = data.copy()
-    expire = datetime.utcnow() + expires_delta
-
-    to_encode.update({"exp": expire})
-    encode_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
-    return encode_jwt
