+++ app/api/routes/users.py
@@ -1,18 +1,15 @@
 from app.services.authentication import check_email_is_taken
 from fastapi import APIRouter, Depends, HTTPException, status
 from fastapi.param_functions import Body
-from jose import JWTError, jwt
 from sqlalchemy.orm import Session
 
 from app.api.dependencies.database import get_db
 from app.api.dependencies.authentication import (
-    create_access_token, get_user_by_username
+    create_access_token, get_current_user
 )
-from app.core.config import ALGORITHM, SECRET_KEY
 from app.db.queries import users as users_db
 from app.models.schemas.memos import Memo, MemoCreate
 from app.models.schemas.users import UserInDB, UserInResponse, UserInUpdate
-from app.services.security import oauth2_scheme
 
 router = APIRouter()
 
@@ -22,30 +19,6 @@ def create_memo_for_user(
     user_id: int, memo: MemoCreate, db: Session = Depends(get_db)
 ) -> Memo:
     return users_db.create_memo_for_user(db=db, memo=memo, user_id=user_id)
-
-
-async def get_current_user(
-    token: str = Depends(oauth2_scheme),
-    db: Session = Depends(get_db)
-) -> UserInDB:
-    credentials_exception = HTTPException(
-        status_code=status.HTTP_401_UNAUTHORIZED,
-        detail="Couldn't invalid credentials",
-        headers={"WWW-Authenticate": "Bearer"},
-    )
-    try:
-        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
-        username: str = payload.get("sub")
-        if username is None:
-            raise credentials_exception
-    except JWTError:
-        raise credentials_exception
-
-    user = get_user_by_username(db, username=username)
-    if user is None:
-        raise credentials_exception
-
-    return user
 
 
 @router.get("/me", response_model=UserInResponse)
